Revert "fix(app): move dialogs to ui library (#707)"
This reverts commit 4aae10a03261a8fce3b90ecce69825a18aca9651.
