+++ packages/app/client/components/index.js
@@ -1,2 +1,3 @@
 export { default as Blocked } from './Blocked'
 export { default as UpdateApp } from './UpdateApp'
+export { default as Dialog } from './Dialog'
