+++ packages/app/client/helpers/index.js
@@ -1,2 +1,4 @@
 export { default as useMediaUpdate } from './useMediaUpdate'
 export { default as useNeedUpdate } from './useNeedUpdate'
+export { alert, confirm, prompt } from './dialogs'
+export { getScope, usePage } from './path'
