+++ packages/babel-plugin-startupjs/moduleMap.json
@@ -11,7 +11,6 @@
     "Card": "/components/Card",
     "Collapse": "/components/Collapse",
     "Content": "/components/Content",
-    "Dialog": "/components/Dialog",
     "Div": "/components/Div",
     "Divider": "/components/Divider",
     "DrawerSidebar": "/components/DrawerSidebar",
@@ -66,10 +65,6 @@
     "ThemeContext": "/theming/ThemeContext.js",
     "ThemeProvider": "/theming/ThemeProvider.js",
     "themed": "/theming/themed.js",
-    "uiAppPlugin": "/helpers/uiAppPlugin.js",
-    "alert": "/helpers/dialogs/alert.js",
-    "confirm": "/helpers/dialogs/confirm.js",
-    "helpers": "/helpers/dialogs/helpers.js",
-    "prompt": "/helpers/dialogs/prompt.js"
+    "uiAppPlugin": "/helpers/uiAppPlugin.js"
   }
 }
