+++ packages/ui/components/Breadcrumbs/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { observer } from 'startupjs'
 import PropTypes from 'prop-types'
-import colorToRGBA from '../../helpers/colorToRGBA'
+import { colorToRGBA } from '../../helpers'
 import Link from './../Link'
 import Row from '../Row'
 import Div from '../Div'
