+++ packages/ui/components/Button/index.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { StyleSheet } from 'react-native'
 import { observer, useIsMountedRef } from 'startupjs'
 import PropTypes from 'prop-types'
-import colorToRGBA from '../../helpers/colorToRGBA'
+import { colorToRGBA } from '../../helpers'
 import Icon from '../Icon'
 import Row from '../Row'
 import Div from '../Div'
