--- packages/ui/components/Dialog/Dialog.en.mdx
@@ -1,79 +1,0 @@
-import { alert, confirm, prompt } from '../../helpers/dialogs'
-import Button from '../Button'
-
-# Dialog
-
-Pop-up API: alert, confirm, prompt
-
-## Using alert
-
-```js
-import { alert } from '@startupjs/ui'
-```
-
-It takes 2 parameters:
-- message (required)
-- title
-
-```jsx example
-async function onOpen () {
-  await alert({ message: 'Test message' })
-  console.log('next line')
-}
-
-return pug`
-  Button(onPress=onOpen) Open alert
-`
-```
-
-## Using confirm
-
-```js
-import { confirm } from '@startupjs/ui'
-```
-
-It takes 4 parameters:
-- message (required)
-- title
-- cancelLabel (default - 'Cancel')
-- confirmLabel (default - 'Confirm')
-
-The function is asynchronous, in the response it returns a boolean value
-
-```jsx example
-async function onOpen () {
-  const result = await confirm({
-    message: 'Are you sure you want to do something?'
-  })
-  console.log(result)
-}
-
-return pug`
-  Button(onPress=onOpen) Open confirm
-`
-```
-
-## Using prompt
-
-```js
-import { prompt } from '@startupjs/ui'
-```
-
-It takes 3 parameters:
-- message (required)
-- title
-- errorMessage
-- messageRequired (default - true)
-
-The function is asynchronous, in the response it returns the entered text
-
-```jsx example
-async function onOpen () {
-  const result = await prompt({ message: 'Enter text' })
-  console.log(result)
-}
-
-return pug`
-  Button(onPress=onOpen) Open prompt
-`
-```
