--- packages/ui/components/Dialog/Dialog.ru.mdx
@@ -1,79 +1,0 @@
-import { alert, confirm, prompt } from '../../helpers/dialogs'
-import Button from '../Button'
-
-# Dialog
-
-API для вспывающих окон: alert, confirm, prompt
-
-## Использование alert
-
-```js
-import { alert } from '@startupjs/ui'
-```
-
-Принимает 2 параметра:
-- message (required)
-- title
-
-```jsx example
-async function onOpen () {
-  await alert({ message: 'Тестовое сообщение' })
-  console.log('next line')
-}
-
-return pug`
-  Button(onPress=onOpen) Открыть alert
-`
-```
-
-## Использование confirm
-
-```js
-import { confirm } from '@startupjs/ui'
-```
-
-Принимает 4 параметра:
-- message (required)
-- title
-- cancelLabel (по умолчанию - 'Cancel')
-- confirmLabel (по умолчанию - 'Confirm')
-
-Функция асинхронная, в ответе возвращает булевое значение
-
-```jsx example
-async function onOpen () {
-  const result = await confirm({
-    message: 'Вы действительно хотите что-то сделать?'
-  })
-  console.log(result)
-}
-
-return pug`
-  Button(onPress=onOpen) Открыть confirm
-`
-```
-
-## Использование prompt
-
-```js
-import { prompt } from '@startupjs/ui'
-```
-
-Принимает 3 параметра:
-- message (required)
-- title
-- errorMessage
-- messageRequired (по умолчанию - true)
-
-Функция асинхронная, в ответе возвращает введеный текст
-
-```jsx example
-async function onOpen () {
-  const result = await prompt({ message: 'Введите текст' })
-  console.log(result)
-}
-
-return pug`
-  Button(onPress=onOpen) Открыть prompt
-`
-```
