--- packages/ui/components/Dialog/index.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import { observer } from 'startupjs'
-import Modal from '../Modal'
-import { usePath } from '../../helpers/path'
-
-export default observer(function Dialog () {
-  const [dialog = {}, $dialog] = usePath('dialog')
-
-  return pug`
-    Modal(
-      visible=dialog.visible
-      onChange=()=> $dialog.set('visible', !dialog.visible)
-      ...dialog
-    )= dialog.children
-  `
-})
