+++ packages/ui/components/Tag/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { StyleSheet } from 'react-native'
 import { observer } from 'startupjs'
 import PropTypes from 'prop-types'
-import colorToRGBA from '../../helpers/colorToRGBA'
+import { colorToRGBA } from '../../helpers'
 import Icon from '../Icon'
 import Div from '../Div'
 import Span from '../typography/Span'
