+++ packages/ui/docs/hooksAndHelpers.js
@@ -1,7 +1,5 @@
 /* @asyncImports */
 import { faCode } from '@fortawesome/free-solid-svg-icons'
-import DialogEn from '../components/Dialog/Dialog.en.mdx'
-import DialogRu from '../components/Dialog/Dialog.ru.mdx'
 import useMediaEn from '../hooks/useMedia.en.mdx'
 import useMediaRu from '../hooks/useMedia.ru.mdx'
 
@@ -13,14 +11,6 @@ export default {
   },
   icon: faCode,
   items: {
-    Dialog: {
-      type: 'mdx',
-      title: 'dialogs',
-      component: {
-        en: DialogEn,
-        ru: DialogRu
-      }
-    },
     useMedia: {
       type: 'mdx',
       title: 'useMedia',
