--- packages/ui/helpers/dialogs/helpers.js
@@ -1,4 +1,0 @@
-import { getScope } from '../path'
-
-export const $dialog = getScope('dialog')
-export const dialogOpen = options => $dialog.set({ visible: true, ...options })
