--- packages/ui/helpers/dialogs/index.js
@@ -1,3 +1,0 @@
-export { default as alert } from './alert'
-export { default as confirm } from './confirm'
-export { default as prompt } from './prompt'
