+++ scripts/babel/createModuleMap.js
@@ -15,8 +15,7 @@ const MODULE_DIRS = {
     { source: path.join(PACKAGES_DIR, 'ui/components') },
     { source: path.join(PACKAGES_DIR, 'ui/hooks'), includeFiles: true },
     { source: path.join(PACKAGES_DIR, 'ui/theming'), includeFiles: true },
-    { source: path.join(PACKAGES_DIR, 'ui/helpers/uiAppPlugin.js'), includeFiles: true },
-    { source: path.join(PACKAGES_DIR, 'ui/helpers/dialogs'), includeFiles: true }
+    { source: path.join(PACKAGES_DIR, 'ui/helpers/uiAppPlugin.js'), includeFiles: true }
   ]
 }
